fix invalid deleting error

+++ api/service/src/main/java/ar/edu/itba/paw/service/LocationServiceImpl.java
@@ -1,9 +1,12 @@
 package ar.edu.itba.paw.service;
 
 import ar.edu.itba.paw.interfaces.dao.LocationDao;
+import ar.edu.itba.paw.interfaces.service.ClinicService;
 import ar.edu.itba.paw.interfaces.service.LocationService;
+import ar.edu.itba.paw.model.Clinic;
 import ar.edu.itba.paw.model.Location;
 import ar.edu.itba.paw.model.exceptions.DuplicateEntityException;
+import ar.edu.itba.paw.model.exceptions.EntityDependencyException;
 import ar.edu.itba.paw.model.exceptions.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -17,6 +20,9 @@ public class LocationServiceImpl implements LocationService {
 
     @Autowired
     private LocationDao locationDao;
+
+    @Autowired
+    private ClinicService clinicService;
 
     @Transactional
     @Override
@@ -55,9 +61,11 @@ public class LocationServiceImpl implements LocationService {
 
     @Transactional
     @Override
-    public long deleteLocation(String name) throws EntityNotFoundException {
+    public long deleteLocation(String name) throws EntityNotFoundException, EntityDependencyException {
         Location location = getLocationByName(name);
         if (location == null) throw new EntityNotFoundException("location");
+        List<Clinic> clinicsInLocation = clinicService.getClinicsByLocation(location);
+        if (!clinicsInLocation.isEmpty()) throw new EntityDependencyException("clinics");
         return locationDao.deleteLocation(name);
     }
 }
